BILL-89 - move existing session code into new service
