+++ src/utils/index.js
@@ -5,7 +5,6 @@ import { rateLimitSetUp } from "./rateLimitSetUp.js";
 import { axiosMiddleware } from "./axiosSetup.js";
 import { initializeDB } from "./sqliteSetup.js";
 import { displayAsciiBanner } from "./displayAsciiBanner.js";
-import { getSessionData } from "./sessionHelper.js";
 
 export {
   getBuildNumber,
@@ -16,5 +15,4 @@ export {
   axiosMiddleware,
   initializeDB,
   displayAsciiBanner,
-  getSessionData,
 };
